# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "House Renting Service API for Zambia - Find and rent houses across Zambia with secure payments via MTN MoMo and Airtel Money",
        "title": "BondiHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "BondiHub Support",
            "url": "https://bondihub.com",
            "email": "support@bondihub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/houses": {
            "get": {
                "description": "Get a paginated list of houses with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Get houses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "House type filter",
                        "name": "house_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum rent filter",
                        "name": "min_rent",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum rent filter",
                        "name": "max_rent",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of bedrooms filter",
                        "name": "bedrooms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of bathrooms filter",
                        "name": "bathrooms",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Houses retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new house listing for rent (landlords and admins only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Create house",
                "parameters": [
                    {
                        "description": "House creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateHouseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "House created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only landlords can create houses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/houses/{id}/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an image for a house listing (owner or admin only)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Upload house image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "House image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Image uploaded successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid house ID or no image file provided",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - You can only upload images for your own houses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "House not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/images/{imageId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an image from a house listing (owner or admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Delete house image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid image ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - You can only delete images for your own houses",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Image not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateHouseRequest": {
            "type": "object",
            "required": [
                "address",
                "description",
                "house_type",
                "monthly_rent",
                "title"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "minLength": 10
                },
                "area": {
                    "type": "number",
                    "minimum": 0
                },
                "bathrooms": {
                    "type": "integer",
                    "minimum": 0
                },
                "bedrooms": {
                    "type": "integer",
                    "minimum": 0
                },
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "house_type": {
                    "type": "string",
                    "enum": [
                        "apartment",
                        "house",
                        "studio",
                        "townhouse",
                        "commercial"
                    ]
                },
                "is_featured": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "monthly_rent": {
                    "type": "number",
                    "minimum": 0
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}